Randomly move a single actor in 3d space
The x3dom data visualization now fetches an arbitrary
length array of position data to plot points in space,
decoupling the generation of swarm data from human
readable visualization.

Eventually individual actor behavior will be governed
by a set of randomly generated parameters, and scored
according to the swarms ability to avoid collision and
survive predators. This should allow for the application
of machine learning techniques to rapidly evolve a "good"
swarm. The end result however, is intended to be viewed
by humans and be reminiscent of fish, bird, and insect
swarms. So, given my deficiency in JS, this was a
necessary first step.