Set default values for an interesting swarm visual
Through experimentation it was discovered that too
many followers cause the leader's pull to be insignificant.
The swarm has a mind of it's own.

At some point the naive `FindClosestN` implementation is
going to be too slow and a better geometric data structure
will be needed.

Followers need to take the swarms boundary box into consideration.

Two swarms, one predator and one prey, would be very interesting.